LayoutTests/css3/masking/clip-path-hit-test-img-expected.txt

Test clip-path is accounted for during hit-testing with replaced elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.elementFromPoint(50, 50).localName is "body"
PASS successfullyParsed is true

TEST COMPLETE
